requirements.txt
flask
python-dotenv
llama-index-core
llama-parse
llama-index-readers-file
word2txt
openpyxl
